+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/utils.py
@@ -718,3 +718,62 @@ def find_matching_prim_paths(prim_path_regex: str, stage: Usd.Stage | None = Non
     for prim in output_prims:
         output_prim_paths.append(prim.GetPath().pathString)
     return output_prim_paths
+
+
+def find_global_fixed_joint_prim(
+    prim_path: str | Sdf.Path, check_enabled_only: bool = False, stage: Usd.Stage | None = None
+) -> UsdPhysics.Joint | None:
+    """Find the fixed joint prim under the specified prim path that connects the target to the simulation world.
+
+    A joint is a connection between two bodies. A fixed joint is a joint that does not allow relative motion
+    between the two bodies. When a fixed joint has only one target body, it is considered to attach the body
+    to the simulation world.
+
+    This function finds the fixed joint prim that has only one target under the specified prim path. If no such
+    fixed joint prim exists, it returns None.
+
+    Args:
+        prim_path: The prim path to search for the fixed joint prim.
+        check_enabled_only: Whether to consider only enabled fixed joints. Defaults to False.
+            If False, then all joints (enabled or disabled) are considered.
+        stage: The stage where the prim exists. Defaults to None, in which case the current stage is used.
+
+    Returns:
+        The fixed joint prim that has only one target. If no such fixed joint prim exists, it returns None.
+
+    Raises:
+        ValueError: If the prim path is not global (i.e: does not start with '/').
+        ValueError: If the prim path does not exist on the stage.
+    """
+    # check prim path is global
+    if not prim_path.startswith("/"):
+        raise ValueError(f"Prim path '{prim_path}' is not global. It must start with '/'.")
+    # get current stage
+    if stage is None:
+        stage = stage_utils.get_current_stage()
+
+    # check if prim exists
+    prim = stage.GetPrimAtPath(prim_path)
+    if not prim.IsValid():
+        raise ValueError(f"Prim at path '{prim_path}' is not valid.")
+
+    fixed_joint_prim = None
+    # we check all joints under the root prim and classify the asset as fixed base if there exists
+    # a fixed joint that has only one target (i.e. the root link).
+    for prim in Usd.PrimRange(prim):
+        # note: ideally checking if it is FixedJoint would have been enough, but some assets use "Joint" as the
+        # schema name which makes it difficult to distinguish between the two.
+        joint_prim = UsdPhysics.Joint(prim)
+        if joint_prim:
+            # if check_enabled_only is True, we only consider enabled joints
+            if check_enabled_only and not joint_prim.GetJointEnabledAttr().Get():
+                continue
+            # check body 0 and body 1 exist
+            body_0_exist = joint_prim.GetBody0Rel().GetTargets() != []
+            body_1_exist = joint_prim.GetBody1Rel().GetTargets() != []
+            # if either body 0 or body 1 does not exist, we have a fixed joint that connects to the world
+            if not (body_0_exist and body_1_exist):
+                fixed_joint_prim = joint_prim
+                break
+
+    return fixed_joint_prim